requirements.txt
fastapi==0.85.0
uvicorn==0.18.3
sqlalchemy==1.4.41
pymysql==1.0.2
aiomysql==0.1.1
jinja2==3.1.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
authlib==1.1.0
starlette==0.20.4
itsdangerous==2.1.2
httpx==0.23.0
databases==0.6.1
aioredis==2.0.1
celery==5.2.7
flower==1.2.0
PyYaml==6.0
redis==4.3.4
email-validator==1.3.0
